third_party/blink/web_tests/external/wpt/dom/events/EventTarget-constructible.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] A constructed EventTarget implements dispatch correctly
  assert_equals: expected null but got object "[object EventTarget]"
Harness: the test ran to completion.
